Change react classes to hooks
